inline NewSearchInputs
Now that NewSearchInputs is only used in client.Plan, we can inline it.
